fix: move declarations to global.d.ts
